updated to snapshot save command adding stopping and starting the vps and checking to make sure the system supports zfs

+++ app/Command/SnapshotCommand/SaveCommand.php
@@ -37,10 +37,16 @@ class SaveCommand extends Command {
             Vps::getLogger()->error("The VPS '{$vzid}' you specified does not appear to exist, check the name and try again.");
             return 1;
         }
-        Vps::getLogger()->error("Creating vz/{$vzid}@first snapshot");
-        Vps::getLogger()->write(Vps::runCommand("zfs destroy vz/{$vzid}@third"));
-        Vps::getLogger()->write(Vps::runCommand("zfs rename vz/{$vzid}@second vz/{$vzid}@third"));
-        Vps::getLogger()->write(Vps::runCommand("zfs rename vz/{$vzid}@first vz/{$vzid}@second"));
-        Vps::getLogger()->write(Vps::runCommand("zfs snapshot vz/{$vzid}@first"));
+        if (Vps::getPoolType() == 'zfs') {
+            Vps::stopVps($vzid);
+            Vps::getLogger()->error("Creating vz/{$vzid}@first snapshot");
+            Vps::getLogger()->write(Vps::runCommand("zfs destroy vz/{$vzid}@third"));
+            Vps::getLogger()->write(Vps::runCommand("zfs rename vz/{$vzid}@second vz/{$vzid}@third"));
+            Vps::getLogger()->write(Vps::runCommand("zfs rename vz/{$vzid}@first vz/{$vzid}@second"));
+            Vps::getLogger()->write(Vps::runCommand("zfs snapshot vz/{$vzid}@first"));
+            Vps::startVps($vzid);
+        } else {
+            Vps::getLogger()->error("This system is not setup for zfs");
+        }
 	}
 }
